error state handling process 2
