+++ android/src/main/java/io/parity/signer/screens/error/ErrorStateScreenGeneral.kt
@@ -74,11 +74,6 @@ internal fun ErrorStateScreen(
 				)
 			}
 		}
-//		PrimaryButtonWide( todo dmitry remove?
-//			label = stringResource(R.string.button_next),
-//			modifier = Modifier.padding(24.dp),
-//			onClicked = {},
-//		)
 	}
 }
